# Improve performance Enumerable.Min()