#38392 

If first nonull element is NaN, return it.

Here is the benchmark code, MinPerf() method is the code after optimized.
```
    public class MinTest
    {
        private float?[] data;

        [Params(2, 10, 100, 1000)]
        public int Length;

        [GlobalSetup]
        public void Setup()
        {
            Random random = new Random();
            this.data = new float?[this.Length];
            this.data[0] = null;
            this.data[1] = float.NaN;
            for (int i = 2; i < this.data.Length; i++)
                this.data[i] = random.Next();
        }

        [Benchmark(Baseline = true)]
        public float? Min()
        {
            return this.data.Min();
        }

        [Benchmark]
        public float? MinPerf()
        {
            return this.data.MinPerf();
        }
    }
```

And results blow
``` ini

BenchmarkDotNet=v0.11.5, OS=Windows 10.0.17134.765 (1803/April2018Update/Redstone4)
Intel Core i7-7500U CPU 2.70GHz (Kaby Lake), 1 CPU, 4 logical and 2 physical cores
Frequency=2835941 Hz, Resolution=352.6166 ns, Timer=TSC
  [Host]     : .NET Framework 4.7.2 (CLR 4.0.30319.42000), 32bit LegacyJIT-v4.7.3416.0
  DefaultJob : .NET Framework 4.7.2 (CLR 4.0.30319.42000), 32bit LegacyJIT-v4.7.3416.0


```
|  Method | Length |          Mean |        Error |       StdDev |        Median | Ratio | RatioSD |
|-------- |------- |--------------:|-------------:|-------------:|--------------:|------:|--------:|
|     **Min** |      **2** |      **51.16 ns** |     **1.314 ns** |     **3.875 ns** |      **51.87 ns** |  **1.00** |    **0.00** |
| MinPerf |      2 |      43.92 ns |     1.996 ns |     5.885 ns |      46.76 ns |  0.87 |    0.15 |
|         |        |               |              |              |               |       |         |
|     **Min** |     **10** |   **1,232.71 ns** |    **16.672 ns** |    **15.595 ns** |   **1,234.14 ns** |  **1.00** |    **0.00** |
| MinPerf |     10 |      45.16 ns |     1.758 ns |     5.183 ns |      47.03 ns |  0.04 |    0.00 |
|         |        |               |              |              |               |       |         |
|     **Min** |    **100** |  **13,926.08 ns** |   **276.106 ns** |   **518.594 ns** |  **14,087.45 ns** | **1.000** |    **0.00** |
| MinPerf |    100 |      47.98 ns |     1.695 ns |     4.998 ns |      48.71 ns | 0.003 |    0.00 |
|         |        |               |              |              |               |       |         |
|     **Min** |   **1000** | **148,046.00 ns** | **2,903.423 ns** | **2,981.601 ns** | **148,040.84 ns** | **1.000** |    **0.00** |
| MinPerf |   1000 |      48.51 ns |     1.174 ns |     3.462 ns |      48.95 ns | 0.000 |    0.00 |