@Wraith2 2019-06-10
Can you provide some benchmarks with different iteration counts to show that this is an improvement in the target case and does not regress the existing performance?