@maryamariyan 2019-06-10
@timandy here's more information for adding benchmarks: https://github.com/dotnet/performance/blob/master/docs/benchmarking-workflow-corefx.md#corefx-prerequisites
While preparing the benchmarks please make sure to build in release and use corerun to run the banchmarks