@Wraith2 2019-06-10
I looked for some in https://github.com/dotnet/performance and didn't find any but it might be worth adding them there so there's an ongoing measure of performance in future