@stephentoub 2019-06-11
It's easy to see how this would improve the performance of a case where there's a NaN at the beginning and then never after that, as in that case it bails immediately rather than having to enumerate the entire sequence.  The potential for concern would be on the worst case, where there's a single non-NaN element in the list: there, we're adding an extra check and branch for no value and the rest of the processing is as small as possible.  I just tried that out, and the extra check made no measurable impact.  While the other checks in the loop are necessary for correctness and this one is not, since it could have a positive impact on some (rare?) cases, no observable negative impact on others, it adds minimal code, and most importantly it adds consistency (every value is now checked for NaN), I'm fine with taking it.